view와의 종속성 제거를 위한 controller 부분의 return 타입 변화.
