+++ src/main/java/com/example/fourart/controller/OAuth2Controller.java
@@ -10,6 +10,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
 import org.springframework.http.*;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.oauth2.core.user.OAuth2User;
@@ -38,15 +39,14 @@ public class OAuth2Controller {
 
     private final InstaConnectService instaConnectService;
     private final MemberService memberService;
-    @Value("${custom.oauth2.instagram.client-id}") String instaClientId;
-    @Value("${custom.oauth2.instagram.client-secret}") String instaClientSecret;
+
     @GetMapping({"","/"})
     public String getAuthorizationMessage(){
         return "Landing_page";
     }
 
     @GetMapping({"/loginSuccess","/success"})
-    public String loginSuccess(@AuthenticationPrincipal OAuth2User oAuth2User){
+    public Member loginSuccess(@AuthenticationPrincipal OAuth2User oAuth2User){
         //여기서 그냥 MemberService 호출해서 Repository 연결하면 될
         /**
          * profile 어트리뷰트 가져오는 것으로 카카오,(구글,네이버 구분)
@@ -87,78 +87,10 @@ public class OAuth2Controller {
         member.setInstagram("");
         member.setInterestingSubject(new ArrayList<>());
         memberService.join(member);
-        return "sign_up";
+        return member;
     }
 
-    @GetMapping("/insta_request")
-    public String instaRequest(){
-        String BASE_URL = "https://api.instagram.com/oauth/authorize?";
-        String clientId=instaClientId;
-        String redirectUri="https://localhost:8443/insta_conn_success";
-        String scope = "user_profile";
-        String responseType = "code";
-        return "redirect:" + BASE_URL + "client_id="+clientId +"&" + "redirect_uri=" + redirectUri +"&scope="+scope +"&response_type="+responseType;
-    }
-    @GetMapping("/insta_conn_success")
-    public String instagramConnect(@RequestParam String code) throws IOException {
 
-        RestTemplate restTemplate = new RestTemplate();
-        HttpHeaders headers = new HttpHeaders();
-        headers.add("Content-Type","application/x-www-form-urlencoded; charset=UTF-8");
-        MultiValueMap<String, String> params = new LinkedMultiValueMap<>();
-
-        params.add("client_id",instaClientId);
-        params.add("client_secret",instaClientSecret);
-        params.add("code",code);
-        params.add("grant_type","authorization_code");
-        params.add("redirect_uri","https://localhost:8443/insta_conn_success");
-
-        HttpEntity<MultiValueMap<String,String>> request = new HttpEntity<>(params,headers);
-
-        ResponseEntity response = restTemplate.exchange(
-                "https://api.instagram.com/oauth/access_token",
-                HttpMethod.POST,
-                request,
-                String.class
-        );
-        ObjectMapper objectMapper = new ObjectMapper();
-
-        InstaOAuth2Token oToken = null;
-
-        try {
-            oToken = objectMapper.readValue(response.getBody().toString(), InstaOAuth2Token.class);
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-
-        return "redirect:getInstaUsername?user_id="+oToken.getUser_id()+"&access_token="+oToken.getAccess_token();
-    }
-    @GetMapping("/getInstaUsername")
-    public String getInstaUsername(@RequestParam("user_id") Long user_id,@RequestParam("access_token") String accessToken,Model model) throws IOException {
-
-        String sUrl = "https://graph.instagram.com/me?fields=id,username&access_token=";
-        URL url = new URL(sUrl+accessToken);
-        HttpsURLConnection con = (HttpsURLConnection) url.openConnection();
-        con.setRequestMethod("GET");
-        con.setRequestProperty("User-Agent", USER_AGENT); // add request header
-        int responseCode = con.getResponseCode();
-        BufferedReader in = new BufferedReader(new InputStreamReader(con.getInputStream()));
-        String inputLine;
-        StringBuffer response = new StringBuffer();
-
-        while ((inputLine = in.readLine()) != null) {
-            response.append(inputLine);
-        }
-        in.close();
-        String result = response.toString();
-        String[] userInfo = result.split(",");
-        String tempUserName = userInfo[1];
-        String userName = tempUserName.split(":")[1];
-        userName = userName.replace("\"","");
-        userName = userName.substring(0,userName.length()-1);
-
-        return "editmyprofile";
-    }
     @GetMapping("/loginFail")
     public String loginFail(){
         return "Landing_page";
