+++ src/main/java/com/example/fourart/service/MemberService.java
@@ -3,12 +3,14 @@ package com.example.fourart.service;
 import com.example.fourart.entity.Member;
 import com.example.fourart.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
 @Service
+@Slf4j
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
 public class MemberService {
@@ -22,13 +24,16 @@ public class MemberService {
     }
 
     private void validateDuplicateMember(Member member){
-        List<Member> findMembers = memberRepository.findByEmail(member.getEmail());
+        List<Member> findMembers = memberRepository.findAllByEmail(member.getEmail());
         if(!findMembers.isEmpty()){
             throw new IllegalStateException("이미 가입된 회원입니다.");
         }
     }
-    public List<Member> findMembers(){
-        return memberRepository.findAll();
+    public Member findMembers(String email){
+        log.info("aaaaaaaaaa");
+        log.info(email);
+        return memberRepository.findByEmail(email);
+
     }
 
 }
